Compare two assays

This plot compares the quantitative proteomics data between two assays. It can be used for example to compare the effect of the normalization process.

The comparison is made with the division operator.

plotCompareAssays(
  assay1,
  assay2,
  conds,
  info = NULL,
  pal.name = NULL,
  subset.view = NULL,
  n = 100,
  type = "scatter",
  FUN = NULL
)

Arguments

assay1

A numeric matrix containing quantitative data before normalization.

assay2

A numeric matrix containing quantitative data after normalization.

conds

A vector of the conditions (one condition per sample).

info

xxx

pal.name

xxx

subset.view

xxx

n

xxx

type

scatter or line

FUN

xxx

Value

A plot

Author

Samuel Wieczorek, Enora Fremy

Examples

data(ft)
assay1 <- assay(ft, 1)
#> Error in assay(ft, 1): could not find function "assay"
assay2 <- 3 * assay(ft, 1)
#> Error in assay(ft, 1): could not find function "assay"
conds <- colData(ft)$Condition
#> Error in colData(ft): could not find function "colData"

## -----------------------------------------------------------------
##
## -----------------------------------------------------------------

plotCompareAssays(assay1, assay2, conds, n = 5)
#> Error in stopifnot(inherits(assay1, "matrix")): object 'assay1' not found