Computes the connex components of a graph, represented by an adjacency matrix.

This function computes the connex components of a graph, represented by an adjacency matrix.

Method to build the list of connected components of an adjacency matrix

Builds a graph as defined by its nodes and edges.

Displays a connected component as a graph with the package 'visNetwork'.

Jitter plot of CC

Display a jitter plot for connex component with the package highcharter

A shiny Module.

ComputeConnectedComposants(obj.se)

get.pep.prot.cc(X)

buildGraph(The.CC, X)

display.CC.visNet(g, obj = NULL, prot.tooltip = NULL, pept.tooltip = NULL)

plotJitterCC(list.of.cc)

plotJitter_hc(df, clickFunction = NULL)

mod_plots_cc_ui(id)

mod_plots_cc_server(input, output, session, cc, matAdj, obj, settings)

Arguments

obj.se

xxxx

X

An adjacency matrix as given by the function BuildListAdjacencyMatrices

The.CC

A cc (a list)

g

A connex component (a list)

obj

xxx

prot.tooltip

A vector of strings which are the text for the tooltips of the proteins.

pept.tooltip

A vector of strings which are the text for the tooltips of the peptides.

list.of.cc

List of cc such as returned by the function get.pep.prot.cc

df

A dataframe containing xxx

clickFunction

A JavaScript code to show an information when a point is clicked

id

shiny id

input

internal

output

internal

session

internal

cc

xxx

matAdj

xxx

Value

A list of items containing the connex components derived from the matrices given in arguments A list of Connected Components (CC) A list of two itens: nodes and edges A plot A plot A plot NA

Author

Samuel Wieczorek

Thomas Burger, Samuel Wieczorek

Thomas Burger, Samuel Wieczorek

Thomas Burger, Samuel Wieczorek

Thomas Burger

Thomas Burger, Samuel Wieczorek

Examples

data(ft)
cc <- ComputeConnectedComposants(ft[[1]])
#> Error: $ operator not defined for this S4 class

data(ft)
X <- adjacencyMatrix(ft[[1]])
#> Error in adjacencyMatrix(ft[[1]]): could not find function "adjacencyMatrix"
ll <- get.pep.prot.cc(X)
#> Error in get.pep.prot.cc(X): object 'X' not found

data(ft)
X <- adjacencyMatrix(ft[[1]])
#> Error in adjacencyMatrix(ft[[1]]): could not find function "adjacencyMatrix"
ll <- get.pep.prot.cc(X)
#> Error in get.pep.prot.cc(X): object 'X' not found
g <- buildGraph(ll[[1]], X)
#> Error in as.matrix(X[The.CC$peptides, The.CC$proteins]): object 'X' not found

data(ft)
X <- adjacencyMatrix(ft[[1]])
#> Error in adjacencyMatrix(ft[[1]]): could not find function "adjacencyMatrix"
ll <- get.pep.prot.cc(X)
#> Error in get.pep.prot.cc(X): object 'X' not found
g <- buildGraph(ll[[1]], X)
#> Error in as.matrix(X[The.CC$peptides, The.CC$proteins]): object 'X' not found
display.CC.visNet(g)
#> Error in visNetwork::visNetwork(g$nodes, g$edges, width = "100%", height = "100%"): object 'g' not found

data(ft)
X <- adjacenyMatrix(ft[[1]])
#> Error in adjacenyMatrix(ft[[1]]): could not find function "adjacenyMatrix"
ll <- get.pep.prot.cc(X)
#> Error in get.pep.prot.cc(X): object 'X' not found
plotJitterCC(ll)
#> Error in stopifnot(!is.null(list.of.cc)): object 'll' not found

df <- data.frame(x=seq_len(10), y=seq_len(10))
plotJitter_hc(df)

xxx
#> Error in eval(expr, envir, enclos): object 'xxx' not found