Bar plot of missing values per lines using highcharter.

This method plots a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins).

listPlotModules()

mod_all_plots_ui(id)

mod_all_plots_server(id, object)

mod_ds_mv_ui(id)

mod_ds_mv_server(id, object, conds, pal.name = reactive({
    NULL
}))

mod_plot_mvImputation_ui(id)

mod_plot_mvImputation_server(
  id,
  object = reactive({     NULL }),
  conds = reactive({     NULL }),
  title = reactive({     NULL }),
  pal.name = reactive({     NULL })
)

mod_ds_variance_ui(id)

mod_ds_variance_server(id, object, conds, pal.name = NULL)

CVDist(object, conds = NULL, pal.name = NULL)

mvPerLinesHisto(qData)

mvPerLinesHistoPerCondition(qData, conds, pal.name)

mvHisto(qData, conds, showValues = FALSE, pal.name = NULL)

Arguments

id

shiny id

object

xxx

conds

A character() representings the condition for each sample of the QFeatures object.

pal.name

A character(1) which is the name of the palette (from the package RColorBrewer to use.

title

A character(1) which is

qData

A data.frame or matrix that contains the quantitative data.

base_palette

xxx

Value

A plot

Details

distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Variance : Builds a densityplot of the CV of entities in numeric matrix. The CV is calculated for each condition present in the dataset (see the slot 'Condition' in the colData() DataFrame)

Heatmap:

The function heatmapD()

The function [] is inspired from the function 'heatmap.2' that displays a numeric matrix. For more information, please refer to the help of the heatmap.2 function.

Missing values

#' - distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Author

Samuel Wieczorek, Enora Fremy

Examples