Bar plot of missing values per lines using highcharter.

This method plots a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins).

listPlotModules()

mod_ds_ui(id)

mod_ds_server(id, object)

Arguments

id

A character(1) for the 'id' of the shiny module. It must be the same as for the '*_ui' function.

object

A instance of the class QFeatures.

Value

A plot

Details

distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Variance : Builds a densityplot of the CV of entities in numeric matrix. The CV is calculated for each condition present in the dataset (see the slot 'Condition' in the colData() DataFrame)

Heatmap:

The function heatmapD()

The function [] is inspired from the function 'heatmap.2' that displays a numeric matrix. For more information, please refer to the help of the heatmap.2 function.

Missing values

#' - distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Author

Samuel Wieczorek, Enora Fremy

Examples

data(ft)
data(ft_na)

qData <- assay(ft, 1)
#> Error in assay(ft, 1): could not find function "assay"
conds <- design(ft)$Condition
#> Error in design(ft): could not find function "design"
pal <- 'Dark2'

data(ft)
qData <- assay(ft, 1)
#> Error in assay(ft, 1): could not find function "assay"
conds <- design(ft)$Condition
#> Error in design(ft): could not find function "design"
densityPlot(qData, conds)
#> Error in densityPlot(qData, conds): object 'conds' not found

legend <- design(ft)$Sample.name
#> Error in design(ft): could not find function "design"
densityPlot(qData, conds, pal.name = 'Dark2')
#> Error in densityPlot(qData, conds, pal.name = "Dark2"): object 'conds' not found

CVDist(qData, conds)
#> Error in stopifnot(inherits(data, "matrix")): object 'qData' not found

#----------------------------------------------
# Plots a heatmap for generic quantitative data
#----------------------------------------------

heatmapD(assay(ft, 1), conds)
#> Error in heatmapD(assay(ft, 1), conds): object 'conds' not found

#----------------------------------------------
# Plots a heatmap for missing values visualization
#----------------------------------------------

#' #----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHisto(qData)
#> Error in stopifnot(inherits(data, "matrix")): object 'qData' not found

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvHisto(qData, conds, pal.name = pal)
#> Error in stopifnot(inherits(data, "matrix")): object 'qData' not found

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

data(ft)
mv.heatmap(assay(ft, 1))
#> Error in assay(ft, 1): could not find function "assay"

#----------------------------------------
# Launch a the shiny module for all plots
# Here, an example with the density plot
#----------------------------------------

if(interactive()){
 data(ft_na)
 ui <- mod_ds_ui('plot')

 server <- function(input, output, session) {
  mod_ds_server('plot', object = reactive({ft_na}))
  }
 
 shinyApp(ui=ui, server=server)
}