Distribution of Observed values with respect to intensity values

This method shows density plots which represents the repartition of Partial Observed Values for each replicate in the dataset. The colors correspond to the different conditions (Condition in colData of the object of class QFeatures). The x-axis represent the mean of intensity for one condition and one entity in the dataset (i. e. a protein) whereas the y-axis count the number of observed values for this entity and the considered condition.

mvPlot.2(qData, conds, title = NULL, pal.name)

mvImage(object, conds)

Arguments

qData

An instance of a matrix containing numeric vales.

conds

A character() of the name of conditions (one condition per sample).

title

The title of the plot

pal.name

A character(1) which is the name of the palette (from the package RColorBrewer to use.

object

An instance of a class SummarizedExperiment.

Value

A plot

xxx

xxxx

missing value image

Plots a heatmap of the quantitative data. Each column represent one of the conditions in the object of class MSnSet and the color is proportional to the mean of intensity for each line of the dataset. The lines have been sorted in order to visualize easily the different number of missing values. A white square is plotted for missing values.

Author

Samuel Wieczorek, Enora Fremy

Examples

library(QFeatures)
data(ft)
qData <- assay(ft,1)
conds <- design(ft)$Condition
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("QFeatures", package = "QFeatures")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("design"), generic = structure("design", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "design"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function 'design' for signature '"QFeatures"'

#-----------------------------
# xxx
#-----------------------------

mvPlot.2(qData, conds)
#> Error in unique(conds): object 'conds' not found

mvPlot.2(Exp1_R25_pept, 2, title="POV distribution", palette=pal)
#> Error in mvPlot.2(Exp1_R25_pept, 2, title = "POV distribution", palette = pal): unused argument (palette = pal)

#-----------------------------
# xxx
#-----------------------------

data(ft_na)
mvImage(ft_na[[1]], conds)
#> Error in mvImage(ft_na[[1]], conds): "qMetadata" %in% colnames(rowData(object)) is not TRUE

#-----------------------------
# xxx
#-----------------------------

if(interactive()){
 library(QFeatures)
 library(shiny)
 library(DaparToolshed)
 data(ft_na)
 conds <- design(ft_na)$Condition
 
  
 ui <- mod_imputation_plot_ui('plot')

 server <- function(input, output, session) {
  mod_imputation_plot_server('plot',
                          obj = reactive({ft_na[[1]]}),
                          conds = reactive({conds}),
                          pal.name = reactive({'Dark2'})
                         )
  }
 
 shinyApp(ui=ui, server=server)
}