Tracking of entities within plots

This shiny module offers a UI to select a subset of a dataset and superimpose quantitative values of this selection on the complete plot Three modes of selection are implemented:

'Protein list': xxx,

'Random': xxx,

'Specific column': xxx

mod_seTracker_ui(id)

mod_seTracker_server(id, se)

Arguments

id

xxx

se

A instance of the class SummarizedExperiment

Examples

if (interactive()){

ui <- tagList(
  mod_seTracker_ui('track'),
  uiOutput('show')
)

server <- function(input, output, session) {
  tmp <- reactiveVal()
  
  tmp <- mod_seTracker_server(id = 'track',
                              se = reactive({ft[[1]]})
  )
  
  output$show <- renderUI({
    p(paste0(tmp(), collapse = ' '))
  })
  
}
shinyApp(ui=ui, server=server)
}