Connected components of a graph.

This function computes the connex components of a graph, represented by its adjacency matrix.

A shiny Module.

wrapper.compute.cc(obj.se)

compute.cc(X)

convertCC2graph(cc, X)

show.graph(g, obj = NULL, prot.tooltip = NULL, pept.tooltip = NULL)

plotJitter_hc(ll.cc, df.tooltips = NULL, clickFunction = NULL)

Extract_CC(ll.cc, type)

mod_explore_graphs_ui(id)

mod_explore_graphs_server(id, ll.cc, se, settings)

mod_plots_cc_ui(id)

mod_plots_cc_server(input, output, session, cc, matAdj, obj, settings)

Arguments

obj.se

An instance of the class SummarizedExperiment. This object must have a adjacencyMatrix.

X

A Matrix which is the adjacency matrix.

cc

xxx

g

xxx

obj

xxx

prot.tooltip

A character() which are the tooltips of the proteins.

pept.tooltip

A character() which are the tooltips of the peptides.

ll.cc

xxx

clickFunction

A JavaScript code to show an information when a point is clicked

type

A character(1) which denotes the type ofconnected components to extract. Available values are: 'OneOne', 'OneMulti' and 'MultiAny'.

id

A character(1) which is the id of the shiny module.

se

An object of class QFeatures

settings

xxx

matAdj

xxx

tooltips

A data.frame which contains a subset of the rowData of the object and will be use as tooltips. Each column of the data.frame will be showed in the global tooltip of a point in the plot.

Value

NA

See also

Author

Thomas Burger, Samuel Wieczorek

Examples

data(ft)
X <- adjacencyMatrix(ft[[1]])
#> Error in adjacencyMatrix(ft[[1]]): could not find function "adjacencyMatrix"

#------------------------------------------
# Build the cc list for an adjaceny matrix
#------------------------------------------

ll.cc <- compute.cc(X)
#> Error in stopifnot(!is.null(X)): object 'X' not found

#-------------------------------------------------
# Convert one cc of this list to a graph structure
#-------------------------------------------------

g <- convertCC2graph(ll.cc[[1]], X)
#> Error in as.matrix(X[cc$peptides, cc$proteins]): object 'X' not found

#-------------------------------------------
# Build the cc list for all type of peptides 
# 'all', 'shared', 'specific'
#-------------------------------------------

ll.cc <- wrapper.compute.cc(ft[[1]])
#> Error: $ operator not defined for this S4 class

#-------------------------------------------------
# Show the graph
#-------------------------------------------------
 
show.graph(g)
#> Error in visNetwork::visNetwork(g$nodes, g$edges, width = "100%", height = "100%"): object 'g' not found

#-----------------------------------------------------
# Plots a graph which x-axis is the number of peptides
# in the cc and the y-axis is the number of proteins
# in the cc.
#'-----------------------------------------------------

tt <- as.data.frame(rowData(ft[[1]])[,1])
#> Error in rowData(ft[[1]]): could not find function "rowData"
colnames(tt) <- colnames(rowData(ft[[1]]))[1]
#> Error in rowData(ft[[1]]): could not find function "rowData"
plotJitter_hc(ll.cc, tt)
#> Error in lapply(ll.cc, function(x) {    length(x$proteins)}): object 'll.cc' not found
plotJitter_hc(ll.cc)
#> Error in lapply(ll.cc, function(x) {    length(x$proteins)}): object 'll.cc' not found

xxx
#> Error in eval(expr, envir, enclos): object 'xxx' not found